+++ template/react-mui-app/src/main.tsx
@@ -2,7 +2,10 @@ import React from "react";
 import { createRoot } from "react-dom/client";
 import { ThemeProvider, CssBaseline } from "@mui/material";
 import App from "./App";
+import { startWebVitalsReporting } from "./performance/report-web-vitals";
 import { appTheme } from "./theme/theme";
+
+startWebVitalsReporting();
 
 createRoot(document.getElementById("root")!).render(
   <React.StrictMode>